reorganize generators
